refactor: move custom email layout to layout dir
